Black some stuff and add a simple token model.
