+++ .github/workflows/django.yml
@@ -39,7 +39,7 @@ jobs:
       - name: Install Dependencies
         run: |
           python -m pip install --upgrade pip
-          pip install -r requirements.txt
+          pip install -r requirements-dev.txt
       - name: Run Tests
         run: |
           python manage.py test --settings main.local_settings
