+++ backends/models.py
@@ -2,8 +2,10 @@
 The models that define our sql tables for the app.
 """
 
+from django.conf import settings
 from django.db import models
 from django.contrib.auth.models import AbstractUser
+
 
 # pylint: disable=W0611, W0107
 class User(AbstractUser):
@@ -17,14 +19,23 @@ class User(AbstractUser):
 class Token(models.Model):
     """
     The backend class for the tokens that allow access to the different backends etc.
-    
+
     Args:
         key: CharField, contains authorization token value.
         user: ForeignKey, foreign key to the logged user.
         created_at: DateTimeField, contains date and time of token creation.
         is_active: BooleanField contains if token is active.
     """
-    
+
+    key = models.CharField(max_length=40, unique=True)
+    user = models.ForeignKey(
+        settings.AUTH_USER_MODEL,
+        on_delete=models.CASCADE,
+    )
+    created_at = models.DateTimeField()
+    is_active = models.BooleanField(default=False)
+
+
 class Backend(models.Model):
     """
     The backend class, which allows us to safe the properties of the backends etc.
