+++ requirements-dev.txt
@@ -1,7 +1,7 @@
 appdirs==1.4.4
 astroid==2.9.3
 attrs==21.1.0
-black==21.5b0
+black==23.1.0
 certifi==2022.12.7
 charset-normalizer==2.0.10
 click==7.1.2
