+++ requirements.txt
@@ -1,6 +1,6 @@
 appdirs==1.4.4
 attrs==21.1.0
-black==21.5b0
+black==23.1.0
 certifi==2022.12.7
 click==7.1.2
 dj-database-url==0.5.0
